requirements.txt
python-telegram-bot==13.8
urllib3==1.26.5
load-dotenv==0.1.0
gspread==6.0.1
oauth2client==4.1.3
google-api-python-client==2.117.0
psycopg2-binary==2.9.9
pygithub==2.3.0 
